+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 <!-- CHANGELOG NEWEST -->
 
-- The worker RxStorage will no langer reuse the worker so that multiple RxDatabase instances can use different workers for better performance.
+- The worker RxStorage will no longer reuse the worker so that multiple RxDatabase instances can use different workers for better performance.
 
 <!-- ADD new changes here! -->